+++ src/main/java/io/jejuwn/controller/ProductInfoController.java
@@ -54,7 +54,7 @@ public class ProductInfoController {
 		
 		ResponseEntity<String> entity = null;
 		try {
-			vo.setId(BigDecimal.valueOf(id));
+			vo.setId(id);
 			service.productInfoUpdate(vo);
 			
 			entity = new ResponseEntity<String>("SUCCESS", HttpStatus.OK);
