+++ src/main/java/io/jejuwn/controller/UserOrderController.java
@@ -54,7 +54,7 @@ public class UserOrderController {
 		
 		ResponseEntity<String> entity = null;
 		try {
-			vo.setId(BigDecimal.valueOf(id));
+			vo.setId(id);
 			service.userOrderUpdate(vo);
 			
 			entity = new ResponseEntity<String>("SUCCESS", HttpStatus.OK);
